Mandatory

+++ libft/ft_strlen.c
@@ -24,7 +24,7 @@
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	size_t	i;
